// src/disaster/enums/observation-type.enum.ts
export enum ObservationType {
  COLLAPSED_BUILDING = 'COLLAPSED_BUILDING',
  DAMAGED_BUILDING = 'DAMAGED_BUILDING',
  ROAD_BLOCKED = 'ROAD_BLOCKED',
  INFRASTRUCTURE_FAILURE = 'INFRASTRUCTURE_FAILURE',
  ASSEMBLY_AREA = 'ASSEMBLY_AREA',
  MEDICAL_POINT = 'MEDICAL_POINT',
  OTHER = 'OTHER',
  FIRE = 'FIRE',
  FIRE_EXTINGUISHED = 'FIRE_EXTINGUISHED',
  GAS_LEAK = 'GAS_LEAK',
  GAS_LEAK_RESOLVED = 'GAS_LEAK_RESOLVED',
  ELECTRICAL_HAZARD = 'ELECTRICAL_HAZARD',
  POWER_ISOLATED = 'POWER_ISOLATED',
  INJURED = 'INJURED',
  INJURED_EVACUATED = 'INJURED_EVACUATED',
  DECEASED = 'DECEASED',
  RESCUE_REQUIRED = 'RESCUE_REQUIRED',
  RESCUE_COMPLETED = 'RESCUE_COMPLETED',
  RESOURCE_NEED = 'RESOURCE_NEED',
  RESOURCE_DISPATCHED = 'RESOURCE_DISPATCHED',
  RESOURCE_DELIVERED = 'RESOURCE_DELIVERED',
  RESOURCE_FULFILLED = 'RESOURCE_FULFILLED',
  DEBRIS_REMOVED = 'DEBRIS_REMOVED',
  STRUCTURE_SECURED = 'STRUCTURE_SECURED',
  ROAD_OPENED = 'ROAD_OPENED',
  SERVICE_RESTORED = 'SERVICE_RESTORED',
}